Telegram Bot/requirements.txt
python-telegram-bot[ext]
yt-dlp
pyzipper
python-dotenv
PySocks
ffmpeg-python
